Move email sending/rendering functions to databaseAPI.
